Collect together patches
