+++ django_plotly_dash/__init__.py
@@ -33,3 +33,4 @@ from .dash_wrapper import DjangoDash
 # Monkeypatching
 
 import django_plotly_dash._callback
+import django_plotly_dash._patches
